+++ engine/internal/process/scope.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/enginekit/config"
 	"github.com/dogmatiq/enginekit/message"
 	"github.com/dogmatiq/testkit/engine/internal/panicx"
 	"github.com/dogmatiq/testkit/envelope"
@@ -19,7 +19,7 @@ import (
 type scope struct {
 	instanceID   string
 	instance     *instance
-	config       configkit.RichProcess
+	config       *config.Process
 	handleMethod string
 	messageIDs   *envelope.MessageIDGenerator
 	observer     fact.Observer
@@ -52,12 +52,12 @@ func (s *scope) End() {
 func (s *scope) ExecuteCommand(m dogma.Command) {
 	mt := message.TypeOf(m)
 
-	if !s.config.MessageTypes()[mt].IsProduced {
+	if !s.config.RouteSet().DirectionOf(mt).Has(config.OutboundDirection) {
 		panic(panicx.UnexpectedBehavior{
 			Handler:        s.config,
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Message:        s.env.Message,
 			Description:    fmt.Sprintf("executed a command of type %s, which is not produced by this handler", mt),
 			Location:       location.OfCall(),
@@ -69,7 +69,7 @@ func (s *scope) ExecuteCommand(m dogma.Command) {
 			Handler:        s.config,
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Message:        s.env.Message,
 			Description:    fmt.Sprintf("executed a command of type %s on an ended process", mt),
 			Location:       location.OfCall(),
@@ -82,7 +82,7 @@ func (s *scope) ExecuteCommand(m dogma.Command) {
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
 			Message:        s.env.Message,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Description:    fmt.Sprintf("executed an invalid %s command: %s", mt, err),
 			Location:       location.OfCall(),
 		})
@@ -94,7 +94,7 @@ func (s *scope) ExecuteCommand(m dogma.Command) {
 		s.now,
 		envelope.Origin{
 			Handler:     s.config,
-			HandlerType: configkit.ProcessHandlerType,
+			HandlerType: config.ProcessHandlerType,
 			InstanceID:  s.instanceID,
 		},
 	)
@@ -117,12 +117,12 @@ func (s *scope) RecordedAt() time.Time {
 func (s *scope) ScheduleTimeout(m dogma.Timeout, t time.Time) {
 	mt := message.TypeOf(m)
 
-	if !s.config.MessageTypes()[mt].IsProduced {
+	if !s.config.RouteSet().DirectionOf(mt).Has(config.OutboundDirection) {
 		panic(panicx.UnexpectedBehavior{
 			Handler:        s.config,
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Message:        s.env.Message,
 			Description:    fmt.Sprintf("scheduled a timeout of type %s, which is not produced by this handler", mt),
 			Location:       location.OfCall(),
@@ -134,7 +134,7 @@ func (s *scope) ScheduleTimeout(m dogma.Timeout, t time.Time) {
 			Handler:        s.config,
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Message:        s.env.Message,
 			Description:    fmt.Sprintf("scheduled a timeout of type %s on an ended process", mt),
 			Location:       location.OfCall(),
@@ -147,7 +147,7 @@ func (s *scope) ScheduleTimeout(m dogma.Timeout, t time.Time) {
 			Interface:      "ProcessMessageHandler",
 			Method:         s.handleMethod,
 			Message:        s.env.Message,
-			Implementation: s.config.Handler(),
+			Implementation: s.config.Source.Get(),
 			Description:    fmt.Sprintf("scheduled an invalid %s timeout: %s", mt, err),
 			Location:       location.OfCall(),
 		})
@@ -160,7 +160,7 @@ func (s *scope) ScheduleTimeout(m dogma.Timeout, t time.Time) {
 		t,
 		envelope.Origin{
 			Handler:     s.config,
-			HandlerType: configkit.ProcessHandlerType,
+			HandlerType: config.ProcessHandlerType,
 			InstanceID:  s.instanceID,
 		},
 	)