+++ envelope/envelope_test.go
@@ -3,8 +3,9 @@ package envelope_test
 import (
 	"time"
 
-	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/enginekit/config"
+	"github.com/dogmatiq/enginekit/config/runtimeconfig"
 	. "github.com/dogmatiq/enginekit/enginetest/stubs"
 	. "github.com/dogmatiq/testkit/envelope"
 	g "github.com/onsi/ginkgo/v2"
@@ -55,7 +56,7 @@ var _ = g.Describe("type Envelope", func() {
 	})
 
 	g.Describe("func NewCommand()", func() {
-		handler := configkit.FromProcess(&ProcessMessageHandlerStub{
+		handler := runtimeconfig.FromProcess(&ProcessMessageHandlerStub{
 			ConfigureFunc: func(c dogma.ProcessConfigurer) {
 				c.Identity("<handler>", "d1c7e18a-4d72-4705-a120-6cfb29eef655")
 				c.Routes(
@@ -73,7 +74,7 @@ var _ = g.Describe("type Envelope", func() {
 			)
 			origin := Origin{
 				Handler:     handler,
-				HandlerType: configkit.ProcessHandlerType,
+				HandlerType: config.ProcessHandlerType,
 				InstanceID:  "<instance>",
 			}
 			now := time.Now()
@@ -98,7 +99,7 @@ var _ = g.Describe("type Envelope", func() {
 	})
 
 	g.Describe("func NewEvent()", func() {
-		handler := configkit.FromAggregate(&AggregateMessageHandlerStub{
+		handler := runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 			ConfigureFunc: func(c dogma.AggregateConfigurer) {
 				c.Identity("<handler>", "8688dc39-b5d0-4468-89fd-0d9452667c0c")
 				c.Routes(
@@ -116,7 +117,7 @@ var _ = g.Describe("type Envelope", func() {
 			)
 			origin := Origin{
 				Handler:     handler,
-				HandlerType: configkit.AggregateHandlerType,
+				HandlerType: config.AggregateHandlerType,
 				InstanceID:  "<instance>",
 			}
 			now := time.Now()
@@ -141,7 +142,7 @@ var _ = g.Describe("type Envelope", func() {
 	})
 
 	g.Describe("func NewTimeout()", func() {
-		handler := configkit.FromProcess(&ProcessMessageHandlerStub{
+		handler := runtimeconfig.FromProcess(&ProcessMessageHandlerStub{
 			ConfigureFunc: func(c dogma.ProcessConfigurer) {
 				c.Identity("<handler>", "1d4e3d22-52fe-4b1b-9bf5-44b2050c08c2")
 				c.Routes(
@@ -160,7 +161,7 @@ var _ = g.Describe("type Envelope", func() {
 			)
 			origin := Origin{
 				Handler:     handler,
-				HandlerType: configkit.ProcessHandlerType,
+				HandlerType: config.ProcessHandlerType,
 				InstanceID:  "<instance>",
 			}
 			now := time.Now()