+++ fact/internal/logging/icon_test.go
@@ -1,7 +1,7 @@
 package logging_test
 
 import (
-	"github.com/dogmatiq/configkit"
+	"github.com/dogmatiq/enginekit/config"
 	. "github.com/dogmatiq/testkit/fact/internal/logging"
 	g "github.com/onsi/ginkgo/v2"
 	gm "github.com/onsi/gomega"
@@ -41,9 +41,9 @@ var _ = g.Describe("func DirectionIcon()", func() {
 
 var _ = g.Describe("func HandlerTypeIcon()", func() {
 	g.It("returns the expected icon", func() {
-		gm.Expect(HandlerTypeIcon(configkit.AggregateHandlerType)).To(gm.Equal(AggregateIcon))
-		gm.Expect(HandlerTypeIcon(configkit.ProcessHandlerType)).To(gm.Equal(ProcessIcon))
-		gm.Expect(HandlerTypeIcon(configkit.IntegrationHandlerType)).To(gm.Equal(IntegrationIcon))
-		gm.Expect(HandlerTypeIcon(configkit.ProjectionHandlerType)).To(gm.Equal(ProjectionIcon))
+		gm.Expect(HandlerTypeIcon(config.AggregateHandlerType)).To(gm.Equal(AggregateIcon))
+		gm.Expect(HandlerTypeIcon(config.ProcessHandlerType)).To(gm.Equal(ProcessIcon))
+		gm.Expect(HandlerTypeIcon(config.IntegrationHandlerType)).To(gm.Equal(IntegrationIcon))
+		gm.Expect(HandlerTypeIcon(config.ProjectionHandlerType)).To(gm.Equal(ProjectionIcon))
 	})
 })